requirements.txt
psycopg2==2.9.3
pyserial==3.5
sqlalchemy==1.4.31
